split sim
